backend/requirements.txt
Flask==2.2.5
Pillow==3.0.10
Flask-CORS==9.5.0